Refactor/shorten stimuli/base.py
